separate traces from incentives
